qa/tasks: move DaemonWatchdog to new file
* Moved DaemonWatchdog class to a new file daemonwatchdog.py
* Dropped the client watch
